+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `unit-conversions` will be documented in this file.
+## 1.1.0 -  2022-11-07
+- add temperature conversion
+- change php version to be ^8.0
+## 1.0.0 - 2022-11-07
+- initial release